switch from coc.nvim to nvim-lsp/nvim-cmp and from chadtree to neo-tree
This removes support for auto-completion from vim invocations, which is
generally fine for me given that I only use vim for minor editing tasks
now.

neo-tree support is still in progress.  Icons are disabled for the time
being.

These changes suggest the need for a more significant refactor to
establish a minimally-suitable vim configuration and a feature-focused
neovim configuration.